DO NOT MERGE !!! CHERRY-PICK Modify the localCache API to require an explicit commit on CachedFileStream.
CachedFileStream has previously performed the commit step in its destructor,
but this means its only recourse for error handling is report_fatal_error.
Modify this to add an explicit commit() method, and call this in the
appropriate places with appropriate error handling for the location.

Currently the destructor of CacheStream calls report_fatal_error if commit()
was not called. This will help track down any remaining uses of the API that
assume the old destructior behaviour. We could instead fall back to the
previous behaviour (and call report_fatal_error only if the commit fails) by
using the commented out code in the desctructor.

Change-Id: Ie3f759ebdd8e5abcd891218aef7785406fb96d3c