[OpenCL][Sema] -fdeclare-opencl-builtins: check if the extensions are enabled, not only their #define
When an OpenCL extension is supported, a #define extname 1 is added.
Then, to use the extension, the developer has to enable it (via the
command line or via a pragma).

Before this patch OpenCL builtins associated to an extension only
checked for macro definitions, and not if the extensions where enabled
or not. This patch fixes this.