[Clang][Builtins] For builtins, use the LangAS mapping for its associated target
This patch enables type checks on several AMDGPU builtins taking
pointers where addressspaces are involved.

In builtin type declarations, we can specify the address space for the
different parameters. For example, v*3 represent 'void addressspace 3'.
However, this doesn't map straight to the target's address space 3. It
goes first through `getLangASForBuiltinAddressSpace`. This function calls
`getLangASFromTargetAS` for all targets except for `AMDGPU`.
For `AMDPGU` it would map 3 to cuda_shared/opencl_local depending of if we're
compiling for CUDA/OpenCL.

In sources with mixed host/device this behaves correctly when compiling
for the device (assuming it is AMDGPU). However, when we compile for
the host (so the target is X86 for example),
`getLangASForBuiltinAddressSpace` maps 3 to addressspace(3); which
is not compatible with cuda_shared/opencl_local. Then type checking
fails.

With this patch, builtins use the `getLangASForBuiltinAddressSpace`
implementation associated for its target.

As a bonus, we do not need anymore the
getCUDABuiltinAddressSpace/getOpenCLBuiltinAddressSpace virtual methods
in ASTContext.